+++ scientific/src/types/scientific.rs
@@ -39,6 +39,7 @@ impl Display for Scientific {
       f.write_char('-')?;
     }
     let exp = self.exponent0();
+    #[allow(clippy::manual_range_contains)]
     if exp >= -1 && exp <= 0 {
       f.write_char('0')?;
       f.write_char('.')?;