+++ scientific/src/types/serde_ser.rs
@@ -1,6 +1,5 @@
 use crate::conversion::bytes_ser::s_to_bytes;
 use crate::types::scientific::Scientific;
-use crate::types::sign::Sign;
 use alloc::string::{String, ToString};
 use alloc::vec::Vec;
 use serde::{Serialize, Serializer};